ADD : switch to integer internal coordinates. Coordinates now vary from -INT_MAX to +INT_MAX rather than -M_PI / +M_PI
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@9612 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ GPS/qgpssatellitetracker.cpp
@@ -42,39 +42,39 @@ QGPSSatelliteTracker::QGPSSatelliteTracker(QWidget *parent) : QWidget(parent)
  *
  * @param QPaintEvent * Ignored by this function
  */
- 
+
 void QGPSSatelliteTracker::paintEvent(QPaintEvent *)
 {
     QPainter painter(this);
-    
+
     painter.setPen(QPen(palette().mid(), 1, Qt::SolidLine));
     painter.setBrush(QBrush(Qt::black, Qt::NoBrush));
-    
+
     // first paint the two reference circles, one at 0 degrees, and
     // the other at 45 degrees
     painter.drawEllipse(rect().x(), rect().y(), rect().width()-1, rect().height()-1);
     painter.drawEllipse(
-        rect().x() + ((rect().width()-1) / 4), 
-        rect().y() + ((rect().height()-1) / 4), 
-        (rect().width()-1) / 2, 
+        rect().x() + ((rect().width()-1) / 4),
+        rect().y() + ((rect().height()-1) / 4),
+        (rect().width()-1) / 2,
         (rect().height()-1) / 2);
-    
+
     // now the reference lines, one vertical and the other horizontal
     painter.drawLine(rect().x(), (rect().height()-1) / 2, rect().width()-1, (rect().height()-1) / 2);
     painter.drawLine((rect().width()-1) / 2, rect().y(), (rect().width()-1) / 2, rect().height()-1);
-    
+
     // now plot the satellites
     painter.setPen(QPen(palette().foreground(), 2, Qt::SolidLine));
     painter.setBrush(QBrush(palette().foreground()));
-    
+
     for(int i = 0; i < 50; i ++)
     {
         int x, y;
-        
+
         if(satArray[i][2] != 0)
         {
             getCoordsFromPos(satArray[i][0], satArray[i][1], x, y);
-            
+
             painter.drawEllipse(x - 2, y - 2, 4, 4);
             painter.drawText(x + 5, y - 1, QString("%1").arg(i));
         }
@@ -94,13 +94,13 @@ void QGPSSatelliteTracker::resetSatInfo()
  *
  * Stores satellite position in the satArray
  */
- 
+
 void QGPSSatelliteTracker::setSatInfo(int index, int elevation, int azimuth, int snr)
 {
     satArray[index][0] = elevation;
     satArray[index][1] = azimuth;
     satArray[index][2] = snr;
-    
+
     update();
 }
 
@@ -116,7 +116,7 @@ void QGPSSatelliteTracker::setSatInfo(int index, int elevation, int azimuth, int
  * @param int &x        Pointer to x-coordinate
  * @param int &y        Pointer to y-coordinate
  */
- 
+
 void QGPSSatelliteTracker::getCoordsFromPos(int elevation, int azimuth, int &x, int &y)
 {
     int theta;
@@ -131,15 +131,15 @@ void QGPSSatelliteTracker::getCoordsFromPos(int elevation, int azimuth, int &x, 
         theta = azimuth - 90;
     else
         theta = 90 - azimuth;
-    
+
     // since the "origin" of the sky is 90 (directly overhead), reverse
     // elevation to make the origin 0 (an elevation of 90 becomes 0, etc)
     elevation = 90 - elevation;
-    
+
     // you should know this (slept too much in trig)
-    x = (cos(theta*M_PI/180)) * elevation;
-    y = (sin(theta*M_PI/180)) * elevation;
-    
+    x = int((cos(theta*M_PI/180)) * elevation);
+    y = int((sin(theta*M_PI/180)) * elevation);
+
     // here we need to make sure that coordinates in quadrants II, III, and IV
     // have the proper signs
     if(azimuth > 270) {
@@ -152,17 +152,17 @@ void QGPSSatelliteTracker::getCoordsFromPos(int elevation, int azimuth, int &x, 
     else if(azimuth > 90) {
         y = y - (2 * y);
     }
-    
+
     // scale our x and y to the size of the widget
-    x = ((double)rect().width() / 180.0) * x;
-    y = ((double)rect().height() / 180.0) * y;
-    
+    x = int(((double)rect().width() / 180.0) * x);
+    y = int(((double)rect().height() / 180.0) * y);
+
     // translate our coordinates by shifting the plane right and up
     // to match the coordinate plane of the widget (to prevent drawing
     // only one quarter of the plane onto the widget
     x = (rect().width() / 2) + x;
     y = (rect().height() / 2) + y;
-    
+
     // mirror y since QPainter's positive-y goes down and the trigonometric
     // positive-y goes up
     y = rect().height() - y;
